requirements.txt
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
matplotlib==3.7.2
yfinance==0.2.18
python-dotenv==1.0.0
newsapi-python==0.2.6
textblob==0.17.1
schedule==1.2.0
flask==2.3.2
pyautogui==0.9.54
psutil==5.9.5